[changed] Adds Nodes component and changes the structure
This commit changes the way the tree is rendered out.  Rather than
recursively using ReactTimber components, ReactTimber is now the
container element.  The Nodes are placed inside.  This lends better
to letting the library handle container level functions rather than have
the users of the library do it.

This commit also adds focusing the root node on initial load. This
contributes to closing part of issue #1