Adds partial down arrow support.
It handles sibling navigation as well as child navigation.
It does not yet handle moving from a child to it's parent's
sibling.  For example,
- A
-- A1
-- A2
- B
It can navigate from A to B if A is collapsed, but in this case
since A is expanded it can only go from A to A1 to A2, but will
not then go to B.